requirements.txt
datasets>=2.14.0
huggingface_hub>=0.20.0
Pillow>=10.0.0
numpy>=1.24.0
anthropic>=0.25.0
openai>=1.14.0
transformers>=4.37.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
python-dotenv>=1.0.0
